rename test_main
